feat: add support for environment variables and prettier configuration feat: add constants for database and response codes feat: add server listener and routes feat: add routes for /populate and /search The changes include adding support for environment variables and prettier configuration. Constants for database and response codes were added. Server listener and routes were added. Routes for /populate and /search were also added.
feat(populate.js, search.js, utils/index.js): add new files and functions
The populate.js file contains a new route that retrieves data from an external source and saves it to a local database. The search.js file contains a new route that searches the local database for data based on specified criteria. The utils/index.js file contains a new function that sends data to the client in a standardized format. These changes add new functionality to the application, allowing users to populate the database with external data and search for data based on specific criteria.